Crate problem062

Problem 62 (Cubic permutations)

Solution detail

This problem can be solved using the following algorithm:

For each range 1...9, 10...99, 100...999, iterate over the cubes in that range.

Create a hash table, with keys given by sets of digits, and values given by the number of cubes having those digits.

After getting to the end of a range, check for keys in the hash table with five cubes, and find the smallest cube corresponding to such an entry.

To speed up the last of these steps, we can store in the hash table not only the number of cubes having a given set of digits, but also the smallest cube having those digits.

Functions

answer
Solve the problem, returning the answer as a String

get_sorted_digits
Get the digits of the given number, in sorted order.

main

solve
Find the smallest cube having exactly the given number of permutations of its digits which are themselves cubes.